README.md
This project is a JavaFX GUI Project dedicated to COVID-19 where this is more or less used to report a law-breakers/offenders to a system(also known as Database). This is not very robust as you might think. This is just a maiden project without considering the MCVE naming conventions.